+++ rate.c
@@ -217,7 +217,7 @@ static void half_band_filter_init(rate_shared_t * p, unsigned which,
 
 typedef struct {
   double     factor;
-  size_t     samples_in, samples_out;
+  uint64_t   samples_in, samples_out;
   int        level, input_stage_num, output_stage_num;
   sox_bool   upsample;
   stage_t    * stages;
@@ -385,8 +385,8 @@ static sample_t const * rate_output(rate_t * p, sample_t * samples, size_t * n)
 static void rate_flush(rate_t * p)
 {
   fifo_t * fifo = &p->stages[p->output_stage_num].fifo;
-  size_t samples_out = p->samples_in / p->factor + .5;
-  size_t remaining = samples_out - p->samples_out;
+  uint64_t samples_out = p->samples_in / p->factor + .5;
+  int64_t remaining = samples_out - p->samples_out;
   sample_t * buff = lsx_calloc(1024, sizeof(*buff));
 
   if ((int)remaining > 0) {
